# Fixed storing of a vaddr as a paddr for DROM flash bank (OCD-910)